@@ -1,46 +1,40 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Reflection;
-using System.Runtime.Loader;
-using System.Threading.Tasks;
-using EventHorizon.Game.Server.Player.Bus;
-using EventHorizon.Game.Server.Player.Controllers;
-using EventHorizon.Game.Server.Player.Performance;
-using EventHorizon.Game.Server.Player.Performance.Impl;
-using EventHorizon.Game.Server.Player.State;
-using EventHorizon.Game.Server.Player.State.Impl;
-using EventHorizon.Game.Server.Player.State.Schedule;
-using EventHorizon.Schedule;
-using IdentityModel.AspNetCore.OAuth2Introspection;
-using MediatR;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.SpaServices.Webpack;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-
-namespace EventHorizon.Game.Server.Player
+﻿namespace EventHorizon.Game.Server.Player
 {
+    using System;
+    using System.Linq;
+    using System.Net;
+    using EventHorizon.Game.Server.Player.Bus;
+    using EventHorizon.Game.Server.Player.Controllers;
+    using EventHorizon.Game.Server.Player.Performance;
+    using EventHorizon.Game.Server.Player.Performance.Impl;
+    using EventHorizon.Game.Server.Player.State;
+    using EventHorizon.TimerService;
+    using MediatR;
+    using Microsoft.AspNetCore.Builder;
+    using Microsoft.AspNetCore.Hosting;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Hosting;
+    using Microsoft.Extensions.Logging;
+
     public class Startup
     {
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
+        public Startup(
+            IConfiguration configuration,
+            IWebHostEnvironment env
+        )
         {
             Configuration = configuration;
             HostingEnvironment = env;
         }
 
         public IConfiguration Configuration { get; }
-        public IHostingEnvironment HostingEnvironment { get; }
+        public IWebHostEnvironment HostingEnvironment { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(
+            IServiceCollection services
+        )
         {
             if (HostingEnvironment.IsDevelopment())
             {
@@ -48,7 +42,6 @@ namespace EventHorizon.Game.Server.Player
                 System.Net.ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12 | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls;
             }
             services.AddHttpClient();
-            services.AddMediatR();
 
             services.AddAuthentication("Bearer")
                 .AddIdentityServerAuthentication(options =>
@@ -57,46 +50,44 @@ namespace EventHorizon.Game.Server.Player
                     options.Authority = Configuration["Auth:Authority"];
                     options.ApiName = Configuration["Auth:ApiName"];
                     options.TokenRetriever = WebSocketTokenRetriever.FromHeaderAndQueryString;
-                    options.JwtBearerEvents.OnMessageReceived = async context =>
-                        {
-                            var accessToken = context.Request.Query["access_token"];
-
-                            if (!string.IsNullOrEmpty(accessToken) &&
-                                (context.HttpContext.WebSockets.IsWebSocketRequest || context.Request.Headers["Accept"] == "text/event-stream"))
-                            {
-                                context.Token = context.Request.Query["access_token"];
-                                context.HttpContext.Request.Headers["Authorization"] = "Bearer " + context.Token;
-                                var result = await context.HttpContext.AuthenticateAsync();
-                                var user = context.HttpContext.User;
-                            }
-                        };
                 });
-            services.AddMvc(options =>
-            {
-                // options.Filters.Add(typeof(JsonExceptionFilter));
-            });
-            services.AddSignalR();
+            services.AddMvc();
+            services.AddSignalR()
+                .AddNewtonsoftJsonProtocol();
             services.AddCors(options => options.AddPolicy("CorsPolicy",
                 builder =>
                 {
-                    builder.AllowAnyMethod().AllowAnyHeader()
-                        .AllowAnyOrigin()
+                    builder.AllowAnyMethod()
+                        .AllowAnyHeader()
+                        .WithOrigins(
+                            Configuration
+                                .GetSection("Cors:Hosts")
+                                .GetChildren()
+                                .AsEnumerable()
+                                .Select(a => a.Value)
+                                .ToArray()
+                        )
                         .AllowCredentials();
                 }));
 
-            services.AddPlayerState(Configuration);
+            services.AddPlayerState(
+                Configuration
+            );
 
+            services.AddTimer();
             services.AddSingleton<IPerformanceTracker, PerformanceTracker>();
 
-            services.AddScheduler((sender, args) =>
-            {
-                Console.WriteLine(args.Exception.Message);
-                args.SetObserved();
-            });
+            services.AddMediatR(
+                typeof(Startup).Assembly
+            );
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(
+            IApplicationBuilder app,
+            IWebHostEnvironment env,
+            ILoggerFactory loggerFactory
+        )
         {
             app.UsePlayerState();
 
@@ -107,12 +98,12 @@ namespace EventHorizon.Game.Server.Player
 
             app.UseCors("CorsPolicy");
             app.UseAuthentication();
+            app.UseAuthorization();
 
-            app.UseSignalR(routes =>
+            app.UseEndpoints(routes =>
             {
                 routes.MapHub<PlayerBus>("/playerBus");
             });
-            app.UseMvc();
         }
     }
 }
