@@ -2,26 +2,34 @@
 using System.Threading.Tasks;
 using Microsoft.Extensions.Hosting;
 
-namespace EventHorizon.Schedule
+namespace EventHorizon.TimerService
 {
     public abstract class HostedService : IHostedService
     {
         private Task _executingTask;
         private CancellationTokenSource _cts;
 
-        public Task StartAsync(CancellationToken cancellationToken)
+        public Task StartAsync(
+            CancellationToken cancellationToken
+        )
         {
             // Create a linked token so we can trigger cancellation outside of this token's cancellation
-            _cts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
+            _cts = CancellationTokenSource.CreateLinkedTokenSource(
+                cancellationToken
+            );
 
             // Store the task we're executing
-            _executingTask = ExecuteAsync(_cts.Token);
+            _executingTask = ExecuteAsync(
+                _cts.Token
+            );
 
             // If the task is completed then return it, otherwise it's running
             return _executingTask.IsCompleted ? _executingTask : Task.CompletedTask;
         }
 
-        public async Task StopAsync(CancellationToken cancellationToken)
+        public async Task StopAsync(
+            CancellationToken cancellationToken
+        )
         {
             // Stop called without start
             if (_executingTask == null)
@@ -33,7 +41,13 @@ namespace EventHorizon.Schedule
             _cts.Cancel();
 
             // Wait until the task completes or the stop token triggers
-            await Task.WhenAny(_executingTask, Task.Delay(-1, cancellationToken));
+            await Task.WhenAny(
+                _executingTask, 
+                Task.Delay(
+                    -1, 
+                    cancellationToken
+                )
+            );
 
             // Throw if cancellation triggered
             cancellationToken.ThrowIfCancellationRequested();
